SW_ThemePark-1/requirements.txt
customtkinter==5.2.0
python-dotenv==1.0.0
Pillow==10.0.0
google-genai==0.5.0
speechrecognition==3.10.0
pyaudio==0.2.13
